candidate route according to new folder structure
