+++ fapi/db/models.py
@@ -1,9 +1,12 @@
+
 from pydantic import BaseModel, EmailStr, Field
 from decimal import Decimal 
 from typing import Optional, List, Literal
 from datetime import time, date, datetime
-from sqlalchemy import Column, Integer, String, DateTime, Date, Boolean
+from sqlalchemy import Column, Integer, String, Enum, DateTime, Boolean, Date ,DECIMAL, Text, ForeignKey, TIMESTAMP
 from sqlalchemy.ext.declarative import declarative_base
+from fapi.db.database import Base
+Base = declarative_base()
 
 
 class UserCreate(BaseModel):
@@ -34,11 +37,11 @@ class UserRegistration(BaseModel):
     registereddate: Optional[str] = None
     level3date: Optional[str] = None
     last: Optional[str] = None
-    visastatus: Optional[str] = Field(None, alias="visaStatus")  
+    visa_status: Optional[str] = Field(None, alias="visaStatus")  
     experience: Optional[str] = None
     education: Optional[str] = None
     specialization: Optional[str] = None
-    referred_by: Optional[str] = Field(None, alias="referredBy")  
+    referby: Optional[str] = Field(None, alias="referredBy")  
 
     class Config:
         allow_population_by_field_name = True
@@ -71,10 +74,17 @@ class ResetPassword(BaseModel):
     new_password: str
 
 
-class GoogleUserCreate(BaseModel):
-    name: str
-    email: str
-    google_id: str
+# --------google_login=-------------
+class AuthUser(Base): 
+    __tablename__ = "authuser"
+
+    id = Column(Integer, primary_key=True, index=True)
+    uname = Column(String(255), unique=True, index=True)
+    fullname = Column(String(255))
+    googleId = Column(String(255))
+    passwd = Column(String(255))
+    status = Column(String(50), default="inactive")
+    registereddate = Column(DateTime, default=datetime.utcnow)
 
 
 # ---------------------------- Innovapath - Request demo --------------------
@@ -108,154 +118,57 @@ class RecentInterview(BaseModel):
     interview_location: Optional[str] = None
     
 # ------------------------------------------- Leads----------------------------------------
+class LeadORM(Base):
+    __tablename__ = "lead"
 
-class LeadBase(BaseModel):
-    full_name: Optional[str] = None
-    entry_date: Optional[datetime] = None
-    phone: Optional[str] = None
-    email: EmailStr
-    workstatus: Optional[str] = None
-    status: Optional[str] = None
-    secondary_email: Optional[str] = None
-    secondary_phone: Optional[str] = None
-    address: Optional[str] = None
-    closed_date: Optional[date] = None
-    notes: Optional[str] = None
-    last_modified: Optional[datetime] = None
-    massemail_unsubscribe: Optional[str] = None
-    massemail_email_sent: Optional[str] = None
-    moved_to_candidate: Optional[bool] = None
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    full_name = Column(String(255))
+    entry_date = Column(DateTime)
+    phone = Column(String(20))
+    email = Column(String(255), nullable=False)
+    workstatus = Column(String(50))
+    status = Column(String(50))
+    secondary_email = Column(String(255))
+    secondary_phone = Column(String(20))
+    address = Column(String(255))
+    closed_date = Column(Date)
+    notes = Column(String(500))
+    last_modified = Column(DateTime)
+    massemail_unsubscribe = Column(String(5))
+    massemail_email_sent = Column(String(5))
+    moved_to_candidate = Column(Boolean)
+
+# -------------------------------------------------------------------------------
+
+class CandidateORM(Base):
+    __tablename__ = "candidate"
+
+    id = Column(Integer, primary_key=True, index=True)
+    full_name = Column(String(100), nullable=True)
+    enrolled_date = Column(Date, nullable=True)
+    email = Column(String(100), nullable=True)
+    phone = Column(String(100), nullable=True)
+    status = Column(String(20), nullable=True)  # No ENUM used
+    workstatus = Column(String(50), nullable=True)  # No ENUM used
+    education = Column(String(200), nullable=True)
+    workexperience = Column(String(200), nullable=True)
+    ssn = Column(String(11), nullable=True)
+    agreement = Column(String(1), default="N", nullable=True)
+    secondaryemail = Column(String(100), nullable=True)
+    secondaryphone = Column(String(45), nullable=True)
+    address = Column(String(300), nullable=True)
+    linkedin_id = Column(String(100), nullable=True)
+    dob = Column(Date, nullable=True)
+    emergcontactname = Column(String(100), nullable=True)
+    emergcontactemail = Column(String(100), nullable=True)
+    emergcontactphone = Column(String(100), nullable=True)
+    emergcontactaddrs = Column(String(300), nullable=True)
+    fee_paid = Column(Integer, nullable=True)
+    notes = Column(Text, nullable=True)
+    batchid = Column(Integer, nullable=False)
 
 
-class LeadCreate(LeadBase):
-    pass
-
-class Lead(LeadBase):
-    id: int
-
-    class Config:
-        from_attributes = True  
-
-
-# class CandidateBase(BaseModel):
-#     name: Optional[str]
-#     enrolleddate: Optional[date]
-#     email: Optional[str]
-#     course: Optional[str]
-#     phone: Optional[str]
-#     status: Optional[str]
-#     workstatus: Optional[str]
-#     education: Optional[str]
-#     workexperience: Optional[str]
-#     ssn: Optional[str]
-#     agreement: Optional[str]
-#     promissory: Optional[str]
-#     driverslicense: Optional[str]
-#     workpermit: Optional[str]
-#     wpexpirationdate: Optional[date]
-#     offerletter: Optional[str]
-#     secondaryemail: Optional[str]
-#     secondaryphone: Optional[str]
-#     address: Optional[str]
-#     city: Optional[str]
-#     state: Optional[str]
-#     country: Optional[str]
-#     zip: Optional[str]
-#     linkedin: Optional[str]
-#     dob: Optional[date]
-#     emergcontactname: Optional[str]
-#     emergcontactemail: Optional[str]
-#     emergcontactphone: Optional[str]
-#     emergcontactaddrs: Optional[str]
-#     guidelines: Optional[str]
-#     ssnvalidated: Optional[str]
-#     bgv: Optional[str]
-#     term: Optional[str]
-#     feepaid: Optional[float]
-#     feedue: Optional[float]
-#     salary0: Optional[str]
-#     salary6: Optional[str]
-#     salary12: Optional[str]
-#     guarantorname: Optional[str]
-#     guarantordesignation: Optional[str]
-#     guarantorcompany: Optional[str]
-#     contracturl: Optional[str]
-#     empagreementurl: Optional[str]
-#     offerletterurl: Optional[str]
-#     dlurl: Optional[str]
-#     workpermiturl: Optional[str]
-#     ssnurl: Optional[str]
-#     referralid: Optional[int]
-#     portalid: Optional[int]
-#     avatarid: Optional[int]
-#     notes: Optional[str]
-#     batchname: str
-#     background: Optional[str]
-#     recruiterassesment: Optional[str]
-#     processflag: Optional[str] = "N"
-#     defaultprocessflag: Optional[str] = "N"
-#     originalresume: Optional[str]
-#     statuschangedate: Optional[date]
-#     diceflag: Optional[str]
-#     batchid: int
-#     emaillist: Optional[str] = "Y"
-#     marketing_startdate: Optional[date]
-#     instructor: Optional[str]
-#     second_instructor: Optional[str]
-
-# class CandidateCreate(CandidateBase):
-#     pass
-
-# class CandidateUpdate(CandidateBase):
-#     pass
-
-# class Candidate(CandidateBase):
-#     candidateid: int
-
-#     class Config:
-#         orm_mode = True
-
-class CandidateBase(BaseModel):
-    full_name: Optional[str]
-    enrolled_date: Optional[date]
-    email: Optional[str]
-    phone: Optional[str]
-    status: Optional[str]
-    workstatus: Optional[str]
-    education: Optional[str]
-    workexperience: Optional[str]
-    ssn: Optional[str]
-    agreement: Optional[str]
-    secondaryemail: Optional[str]
-    secondaryphone: Optional[str]
-    address: Optional[str]
-    linkedin_id: Optional[str]
-    dob: Optional[date]
-    emergcontactname: Optional[str]
-    emergcontactemail: Optional[str]
-    emergcontactphone: Optional[str]
-    emergcontactaddrs: Optional[str]
-    fee_paid: Optional[int]
-    notes: Optional[str]
-    batchid: int
-
-class CandidateCreate(CandidateBase):
-    pass
-
-class CandidateUpdate(CandidateBase):
-    pass
-
-class Candidate(CandidateBase):
-    id: int
-
-    class Config:
-        from_attributes = True
-
-class PaginatedCandidateResponse(BaseModel):
-    page: int
-    limit: int
-    total: int
-    data: List[Candidate]
+        
 
 class PlacementBase(BaseModel):
     candidate_id: Optional[int]
@@ -348,8 +261,6 @@ class UnsubscribeRequest(BaseModel):
 
 
 
-
-
 class VendorBase(BaseModel):
     id: Optional[int]
     full_name: Optional[str]
@@ -378,45 +289,41 @@ class VendorResponse(VendorBase):
 
 
 
-# --------------------------------------Candidate_placement-------------------------------
-class CandidateMarketingBase(BaseModel):
-    candidate_id: int
-    primary_instructor_id: Optional[int] = None
-    sec_instructor_id: Optional[int] = None
-    marketing_manager: Optional[int] = None
-    start_date: date
-    notes: Optional[str] = None
-    status: Literal['active', 'break', 'not responding']
-
-class CandidateMarketingCreate(CandidateMarketingBase):
-    pass
-
-class CandidateMarketing(CandidateMarketingBase):
-    id: int
-    last_mod_datetime: Optional[datetime]
-
-    class Config:
-        from_attributes = True
+# --------------------------------------Candidate_Marketing-------------------------------
 
 
+class CandidateMarketingORM(Base):
+    __tablename__ = "candidate_marketing"
 
-class CandidatePlacementBase(BaseModel):
-    candidate_id: int
-    company: str
-    placement_date: date
-    type: Optional[Literal['Company', 'Client', 'Vendor', 'Implementation Partner']] = None
-    status: Literal['scheduled', 'cancelled']
-    base_salary_offered: Optional[float] = None
-    benefits: Optional[str] = None
-    fee_paid: Optional[float] = None
-    notes: Optional[str] = None
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    # candidate_id = Column(Integer, ForeignKey("candidate.candidateid", ondelete="CASCADE"), nullable=False)
+    candidate_id = Column(Integer)
+    # primary_instructor_id = Column(Integer, ForeignKey("employee.id"), nullable=True)
+    primary_instructor_id = Column(Integer)
+    # sec_instructor_id = Column(Integer, ForeignKey("employee.id"), nullable=True)
+    sec_instructor_id = Column(Integer)
+    # marketing_manager = Column(Integer, ForeignKey("employee.id"), nullable=True)
+    marketing_manager = Column(Integer)
+    start_date = Column(Date, nullable=False)
+    notes = Column(Text, nullable=True)
+    status = Column(Enum('active', 'break', 'not responding'), nullable=False)
+    last_mod_datetime = Column(TIMESTAMP, default=datetime.utcnow, onupdate=datetime.utcnow)
 
-class CandidatePlacementCreate(CandidatePlacementBase):
-    pass
 
-class CandidatePlacement(CandidatePlacementBase):
-    id: int
-    last_mod_datetime: Optional[datetime]
+# --------------------------------------Candidate_Placement-------------------------------
 
-    class Config:
-        from_attributes = True
+class CandidatePlacementORM(Base):
+    __tablename__ = "candidate_placement"
+
+    id = Column(Integer, primary_key=True, autoincrement=True)
+    # candidate_id = Column(Integer, ForeignKey("candidate.candidateid", ondelete="CASCADE"), nullable=False)
+    candidate_id = Column(Integer)
+    company = Column(String(200), nullable=False)
+    placement_date = Column(Date, nullable=False)
+    type = Column(Enum('Company', 'Client', 'Vendor', 'Implementation Partner'), nullable=True)
+    status = Column(Enum('scheduled', 'cancelled'), nullable=False)
+    base_salary_offered = Column(DECIMAL(10, 2), nullable=True)
+    benefits = Column(Text, nullable=True)
+    fee_paid = Column(DECIMAL(10, 2), nullable=True)
+    notes = Column(Text, nullable=True)
+    last_mod_datetime = Column(TIMESTAMP, default=None, onupdate=None)
