--- fapi/models.py
@@ -1,400 +1,0 @@
-# from pydantic import BaseModel, EmailStr, Field
-# from decimal import Decimal 
-# from typing import Optional, List, Literal
-# from datetime import time, date, datetime
-# from sqlalchemy import Column, Integer, String, DateTime, Date, Boolean
-# from sqlalchemy.ext.declarative import declarative_base
-
-
-# class UserCreate(BaseModel):
-#     uname: str
-#     passwd: str
-
-
-
-# class UserRegistration(BaseModel):
-#     uname: str
-#     passwd: str
-#     dailypwd: Optional[str] = None
-#     team: Optional[str] = None
-#     level: Optional[str] = None
-#     instructor: Optional[str] = None
-#     override: Optional[str] = None
-#     status: Optional[str] = None
-#     lastlogin: Optional[str] = None
-#     logincount: Optional[str] = None
-#     # firstname: Optional[str] = None
-#     # lastname: Optional[str] = None
-#     firstname: Optional[str] = Field(None, alias="firstName")
-#     lastname: Optional[str] = Field(None, alias="lastName")
-#     fullname: Optional[str] = None
-#     phone: Optional[str] = None
-#     address: Optional[str] = None
-#     city: Optional[str] = None
-#     Zip: Optional[str] = None
-#     country: Optional[str] = None
-#     message: Optional[str] = None
-#     registereddate: Optional[str] = None
-#     level3date: Optional[str] = None
-#     last: Optional[str] = None
-#     visastatus: Optional[str] = Field(None, alias="visaStatus")  # with alias
-#     experience: Optional[str] = None
-#     education: Optional[str] = None
-#     specialization: Optional[str] = None
-#     referred_by: Optional[str] = Field(None, alias="referredBy")  # with alias
-
-#     class Config:
-#         allow_population_by_field_name = True
-#         allow_population_by_alias = True
-
-
-# class ContactForm(BaseModel):
-#     firstName: str
-#     lastName: str
-#     email: str
-#     phone: str
-#     message: str
-
-
-# class Token(BaseModel):
-#     access_token: str
-#     token_type: str
-
-
-# class EmailRequest(BaseModel):
-#     email: EmailStr
-
-
-# class ResetPasswordRequest(BaseModel):
-#     email: EmailStr
-
-
-# class ResetPassword(BaseModel):
-#     token: str
-#     new_password: str
-
-
-# class GoogleUserCreate(BaseModel):
-#     name: str
-#     email: str
-#     google_id: str
-
-
-# # ---------------------------- Innovapath - Request demo --------------------
-
-# class VendorCreate(BaseModel):
-#     full_name: str
-#     phone_number: str
-#     email: Optional[EmailStr] = None
-#     city: Optional[str] = None
-#     postal_code: Optional[str] = None
-#     address: Optional[str] = None
-#     country: Optional[str] = None
-#     note: Optional[str] = None
-# # --------------------------------------------------------------------------
-
-# class RecentPlacement(BaseModel):
-#     id: int
-#     candidate_name: str
-#     company: str
-#     position: str
-#     placement_date: str
-
-
-# # class RecentInterview(BaseModel):
-# #     id: int
-# #     candidate_name: str
-# #     candidate_role: Optional[str]
-# #     interview_time: time
-# #     interview_date: date
-# #     interview_mode: Optional[str]
-# #     client_name: Optional[str]
-# #     interview_location: Optional[str]
-# #     created_at: datetime
-# class RecentInterview(BaseModel):
-#     candidate_name: str
-#     candidate_role: Optional[str] = None
-#     interview_time: str
-#     interview_date: str
-#     interview_mode: Optional[str] = None
-#     client_name: Optional[str] = None
-#     interview_location: Optional[str] = None
-    
-# # ------------------------------------------- Leads----------------------------------------
-
-
-# from pydantic import BaseModel, EmailStr, Field
-# from typing import Optional
-# from datetime import datetime, date
-
-# class LeadBase(BaseModel):
-#     full_name: Optional[str] = None
-#     entry_date: Optional[datetime] = None
-#     phone: Optional[str] = None
-#     email: EmailStr
-#     workstatus: Optional[str] = None
-#     status: Optional[str] = None
-#     secondary_email: Optional[str] = None
-#     secondary_phone: Optional[str] = None
-#     address: Optional[str] = None
-#     closed_date: Optional[date] = None
-#     notes: Optional[str] = None
-#     last_modified: Optional[datetime] = None
-#     massemail_unsubscribe: Optional[str] = None
-#     massemail_email_sent: Optional[str] = None
-#     moved_to_candidate: Optional[bool] = None
-
-
-# class LeadCreate(LeadBase):
-#     pass
-
-# class Lead(LeadBase):
-#     id: int
-
-#     class Config:
-#         from_attributes = True  
-
-
-# class CandidateBase(BaseModel):
-#     name: Optional[str]
-#     enrolleddate: Optional[date]
-#     email: Optional[str]
-#     course: Optional[str]
-#     phone: Optional[str]
-#     status: Optional[str]
-#     workstatus: Optional[str]
-#     education: Optional[str]
-#     workexperience: Optional[str]
-#     ssn: Optional[str]
-#     agreement: Optional[str]
-#     promissory: Optional[str]
-#     driverslicense: Optional[str]
-#     workpermit: Optional[str]
-#     wpexpirationdate: Optional[date]
-#     offerletter: Optional[str]
-#     secondaryemail: Optional[str]
-#     secondaryphone: Optional[str]
-#     address: Optional[str]
-#     city: Optional[str]
-#     state: Optional[str]
-#     country: Optional[str]
-#     zip: Optional[str]
-#     linkedin: Optional[str]
-#     dob: Optional[date]
-#     emergcontactname: Optional[str]
-#     emergcontactemail: Optional[str]
-#     emergcontactphone: Optional[str]
-#     emergcontactaddrs: Optional[str]
-#     guidelines: Optional[str]
-#     ssnvalidated: Optional[str]
-#     bgv: Optional[str]
-#     term: Optional[str]
-#     feepaid: Optional[float]
-#     feedue: Optional[float]
-#     salary0: Optional[str]
-#     salary6: Optional[str]
-#     salary12: Optional[str]
-#     guarantorname: Optional[str]
-#     guarantordesignation: Optional[str]
-#     guarantorcompany: Optional[str]
-#     contracturl: Optional[str]
-#     empagreementurl: Optional[str]
-#     offerletterurl: Optional[str]
-#     dlurl: Optional[str]
-#     workpermiturl: Optional[str]
-#     ssnurl: Optional[str]
-#     referralid: Optional[int]
-#     portalid: Optional[int]
-#     avatarid: Optional[int]
-#     notes: Optional[str]
-#     batchname: str
-#     background: Optional[str]
-#     recruiterassesment: Optional[str]
-#     processflag: Optional[str] = "N"
-#     defaultprocessflag: Optional[str] = "N"
-#     originalresume: Optional[str]
-#     statuschangedate: Optional[date]
-#     diceflag: Optional[str]
-#     batchid: int
-#     emaillist: Optional[str] = "Y"
-#     marketing_startdate: Optional[date]
-#     instructor: Optional[str]
-#     second_instructor: Optional[str]
-
-# class CandidateCreate(CandidateBase):
-#     pass
-
-# class CandidateUpdate(CandidateBase):
-#     pass
-
-# class Candidate(CandidateBase):
-#     candidateid: int
-
-#     class Config:
-#         orm_mode = True
-
-        
-
-# class PlacementBase(BaseModel):
-#     candidate_id: Optional[int]
-#     candidate_name: Optional[str]
-#     candidate_email: Optional[str]
-#     company: Optional[str]
-#     client_id: Optional[int]
-#     batch: Optional[str]
-#     # placed_date: Optional[date]
-#     placement_date: Optional[date]
-#     status: Optional[str]
-#     marketing_email_address: Optional[str]
-#     vendor_or_client_name: Optional[str]
-#     vendor_or_client_contact: Optional[str]
-#     start_date: Optional[date]
-#     position: Optional[str]
-#     amount_paid: Optional[float]
-#     work_authorization: Optional[str]
-#     experience_in_resume: Optional[str]
-#     role: Optional[str]
-#     job_location: Optional[str]
-#     terms_and_conditions: Optional[str]
-#     notes: Optional[str]
-#     candidate_profile_folder: Optional[str]
-#     placement_verified: Optional[bool] = False
-#     joining_letter_url: Optional[str]
-
-
-# class PlacementCreate(PlacementBase):
-#     pass
-
-
-# class PlacementUpdate(PlacementBase):
-#     pass
-
-
-# class Placement(PlacementBase):
-#     id: int
-
-#     class Config:
-#         orm_mode = True
-# # =======
-
-
-# class RecentPlacement(BaseModel):
-#     id: int
-#     candidate_name: str
-#     company: str
-#     position: str
-#     placement_date: str
-
-
-# class RecentInterview(BaseModel):
-#     id: int
-#     candidate_name: str
-#     candidate_role: Optional[str]
-#     interview_time: time
-#     interview_date: date
-#     interview_mode: Optional[str]
-#     client_name: Optional[str]
-#     interview_location: Optional[str]
-#     created_at: datetime
-
-
-# # .......................................NEW INNOVAPATH..............................
-
-
-# class CandidateMarketingBaseModel(BaseModel):
-#     candidate_id: int
-#     full_name: str
-#     email: str
-#     phone: str
-#     role: str
-#     experience: int
-#     location: str
-#     availability: str
-#     skills: str
-
-# # class CandidateMarketingCreate(CandidateMarketingBase):
-#     # pass
-
-# class CandidateMarketing(CandidateMarketingBaseModel):
-#     id: int
-    
-#     class Config:
-#         orm_mode = True
-
-
-
-
-
-
-
-
-# class VendorBase(BaseModel):
-#     id: Optional[int]
-#     full_name: Optional[str]
-#     phone_number: Optional[str]
-#     email: Optional[str]
-#     city: Optional[str] = None
-#     postal_code: Optional[str] = None
-#     address: Optional[str] = None
-#     country: Optional[str] = None
-#     type: Optional[str] = None
-#     note: Optional[str] = None
-#     last_contacted: Optional[date] = None
-# class VendorCreate(VendorBase):
-#     pass
-
-# class VendorUpdate(VendorBase):
-#     pass
-
-# class VendorResponse(VendorBase):
-#     id: int  # correct usage in Pydantic
-
-#     class Config:
-#         from_attributes = True  # Use this if using Pydantic v2 instead of `orm_mode = True`
-
-
-
-
-
-# # --------------------------------------Candidate_placement-------------------------------
-# class CandidateMarketingBase(BaseModel):
-#     candidate_id: int
-#     primary_instructor_id: Optional[int] = None
-#     sec_instructor_id: Optional[int] = None
-#     marketing_manager: Optional[int] = None
-#     start_date: date
-#     notes: Optional[str] = None
-#     status: Literal['active', 'break', 'not responding']
-
-# class CandidateMarketingCreate(CandidateMarketingBase):
-#     pass
-
-# class CandidateMarketing(CandidateMarketingBase):
-#     id: int
-#     last_mod_datetime: Optional[datetime]
-
-#     class Config:
-#         from_attributes = True
-
-
-
-# class CandidatePlacementBase(BaseModel):
-#     candidate_id: int
-#     company: str
-#     placement_date: date
-#     type: Optional[Literal['Company', 'Client', 'Vendor', 'Implementation Partner']] = None
-#     status: Literal['scheduled', 'cancelled']
-#     base_salary_offered: Optional[float] = None
-#     benefits: Optional[str] = None
-#     fee_paid: Optional[float] = None
-#     notes: Optional[str] = None
-
-# class CandidatePlacementCreate(CandidatePlacementBase):
-#     pass
-
-# class CandidatePlacement(CandidatePlacementBase):
-#     id: int
-#     last_mod_datetime: Optional[datetime]
-
-#     class Config:
-#         from_attributes = True
